+++ lib/src/channel/osm_method_channel.dart
@@ -197,10 +197,24 @@ class MethodChannelOSM extends OSMPlatform {
 
   /// select position and show marker on it
   @override
-  Future<GeoPoint> pickLocation(int idOSM) async {
+  Future<GeoPoint> pickLocation(
+    int idOSM, {
+    GlobalKey? key,
+    String imageURL = "",
+  }) async {
+    Uint8List? bitmap;
+    Map args = {};
+    if (key != null) {
+      bitmap = await _capturePng(key);
+      args.addAll({"icon": bitmap});
+    }
+    if(imageURL.isNotEmpty){
+      args.addAll({"imageURL": imageURL});
+    }
+
     try {
       Map<String, dynamic> map = (await (_channels[idOSM]!
-          .invokeMapMethod("user#pickPosition", null)))!;
+          .invokeMapMethod("user#pickPosition", args)))!;
       return GeoPoint(latitude: map["lat"], longitude: map["lon"]);
     } on PlatformException catch (e) {
       throw GeoPointException(msg: e.message);